Removing PeriodUtils
